requirements.txt
Django==5.0.*
djangorestframework==3.15.*
djangorestframework-simplejwt==5.3.*
djoser==2.2.*
drf_yasg==1.21.11
psycopg2-binary
